Add files via upload
This is the login/signup page of Anime-guru.
So let's start up, So what is Anime-guru and what is it used for?
Anime-guru is an online website where people can watch anime and read Mangas. This is an incomplete website, So I need ur collaboration and help and advices so we could make this work.